refactor(import): move CSV upload form into /transactions/import page
Removes the intermediate ImportCSVDrawer and consolidates the upload
and review steps into a single two-step page. The "Importar transações"
menu item now navigates directly to /transactions/import without opening
a drawer first.
